+++ test/py_test/test_b2b_DP.py
@@ -54,8 +54,19 @@ def verify_result_obj(res_obj):
     assert res_obj.value() == 0
 
 
+def supports_buffer_protocol(obj):
+    """
+    Helper function to test whether the data type supports Python buffer protocol.
+    """
+    try:
+        memoryview(obj)
+        return True
+    except TypeError:
+        return False
+
+
 @pytest.mark.b2b
-def test_b2b_send_bytes_recv_list():
+def test_b2b_send_bytes_recv_bytes():
     """
     Back-to-back test for Segmenter::sendEvent() and Reassembler::getEvent()
     using the py::bytes & List interface.
@@ -86,13 +97,14 @@ def test_b2b_send_bytes_recv_list():
     time.sleep(1)  # has to wait for a few cycles for the reassembler to recv the msg
 
     # Prepare the recv buffer. Python List is mutable. Use list[0] to get the context.
-    recv_bytes_list = [None]
-    res, recv_event_len, recv_event_num, recv_data_id = reas.getEvent(recv_bytes_list)
-    assert res == 0   # the error code
+    recv_len, recv_bytes, recv_event_num, recv_data_id = reas.getEventBytes()
+    assert recv_len >= 0, "Reassembler getEvent error! "   # the received nBytes
+    assert recv_len == len(send_context), "recv_length did not match"
     assert recv_data_id == DATA_ID, "recv_data_id did not match"
-    assert recv_event_len == len(send_context), "recv_length did not match"
-    recv_str = recv_bytes_list[0].decode('utf-8')
+    recv_str = recv_bytes.decode('utf-8')
     assert recv_str == SEND_STR, "recv_str did not match"
+
+    reas.stopThreads()
 
 
 # NOTE: Launch in the main suite will fail. Lauch with -m b2b will succeed.
@@ -124,22 +136,86 @@ def test_b2b_send_numpy_get_numpy():
     time.sleep(1)
 
     # Receive the numpy array
-    res, recv_array, recv_event_num, recv_data_id = reas.get1DNumpyArray(np.float32().dtype)
-    # print(res, recv_array.nbytes, recv_event_num, recv_data_id)
-    assert res == 0   # the error code
+    recv_len, recv_array, recv_event_num, recv_data_id = reas.get1DNumpyArray(np.float32().dtype)
+    # print(recv_bytes, recv_array.nbytes, recv_event_num, recv_data_id)
+    assert recv_len > 0
+    assert recv_len == recv_array.nbytes
     assert recv_data_id == DATA_ID, "recv_data_id did not match"
     assert recv_array.nbytes == send_bytes, "recv_length did not match"
     assert np.array_equal(send_array.flatten(), recv_array), "recv_array did not match"
 
+    reas.stopThreads()
+
+
+@pytest.mark.b2b
+def test_b2b_send_numpy_queue_recv_bytes():
+    """
+    Back-to-back test for Segmenter::addToSendQueue() and Reassembler::recvEvent().
+    Send with numpy and receive with Python bytes.
+    """
+
+    seg = get_segmenter()
+    reas = get_reassembler()
+
+    # Start the reassembler
+    res = reas.OpenAndStart()
+    verify_result_obj(res)
+
+    # Start the segmenter
+    res = seg.OpenAndStart()
+    verify_result_obj(res)
+
+    # Send 1D numpy array 5 times
+    num_elements = 100
+    # Create a 2D numpy array and each time send 1 row
+    send_array = np.array([np.full((num_elements,), i, dtype=np.uint8) for i in range(5)])
+    for i in range(5):
+        send_bytes = send_array[i].nbytes
+        res = seg.addToSendQueue(send_array[i], send_bytes)
+        verify_result_obj(res)
+
+    time.sleep(1)
+
+    # Print the msgCnt of SendStats
+    res = seg.getSendStats()
+    assert res.errCnt == 0, f"Send error occured with {res.msgCnt} msg sent"
+    # print(f"Sent msg count: {res.msgCnt}, sent bytes: {send_array.nbytes}")
+
+    # Recieve via a while loop
+    total_bytes = 0
+    i = sleep_cnt = 0
+    while True:
+        i += 1
+        recv_len, recv_data, recv_event_num, recv_data_id = reas.recvEventBytes(wait_ms=100)
+        if (recv_len == -2 or sleep_cnt > 4 or total_bytes >= send_array.nbytes):
+            # print(f"Receiving error after recv {total_bytes} bytes")
+            break
+
+        if recv_len == -1:
+            sleep_cnt += 1
+            time.sleep(5)
+            continue
+
+        total_bytes += recv_len
+        # Validate the recv buffer
+        # Decode the buffer as numpy 1D array
+        recv_array = np.frombuffer(recv_data, dtype=np.uint8)
+        assert np.array_equal(send_array[i - 1], recv_array)
+        assert recv_data_id == DATA_ID
+
+        if total_bytes >= send_array.nbytes:
+            break
+
+    reas.stopThreads()
+
 
 # @pytest.mark.b2b
 # Launch this one with python istead of pytest -m
-
-# NOTE: It's not included in the b2b test because the numpy array size is really big and
-# may fail.
+#
+# NOTE: It's not included in the b2b test because the numpy array size is big and may fail.
 #
 # A fail run:
-# (e2sar) (e2sar) python test/py_test/test_b2b_DP.py 
+# (e2sar) (e2sar) python test/py_test/test_b2b_DP.py
 # Send numpy array of 200.0 MB for 5 times
 # Sent msg count: 282863, send MB: 1000.0
 # Rd 1, received 200.0 MB, total 200.0 MB
@@ -156,7 +232,7 @@ def test_b2b_send_numpy_get_numpy():
 @pytest.mark.skip(reason="Excluded from main suite")
 def test_b2b_send_numpy_queue_get_numpy():
     """
-    A "bonuns" Back-to-back test for Segmenter::addToSendQueue() and Reassembler::getEvent()
+    A "bonus" Back-to-back test for Segmenter::addToSendQueue() and Reassembler::getEvent()
     with numpy interfaces.
     """
 
@@ -186,27 +262,29 @@ def test_b2b_send_numpy_queue_get_numpy():
     # Print the msgCnt of SendStats
     res = seg.getSendStats()
     assert res.errCnt == 0, f"Send error occured with {res.msgCnt} msg sent"
-    print(f"Sent msg count: {res.msgCnt}, send MB: {send_array.nbytes / 1000000}")
+    print(f"Sent msg count: {res.msgCnt}, sent MB: {send_array.nbytes / 1000000}")
 
     # Recieve via a while loop
-    recv_bytes = 0
-    i = sleep_time = 0
-    while (True):
+    total_bytes = 0
+    i = sleep_cnt = 0
+    while True:
         i += 1
-        res, recv_array, recv_event_num, recv_data_id = reas.get1DNumpyArray(np.int32().dtype)
-        if (res == -2 or sleep_time > 4 or recv_bytes > send_array.nbytes):
-            print(f"Receiving error after recv {recv_bytes} bytes")
+        recv_len, recv_array, recv_event_num, recv_data_id = reas.get1DNumpyArray(np.uint8().dtype)
+        if (recv_len == -2 or sleep_cnt > 4 or total_bytes >= send_array.nbytes):
+            print(f"Receiving error after recv {total_bytes} bytes")
             break
 
-        if res == -1:
-            sleep_time += 1
+        if recv_len == -1:
+            sleep_cnt += 1
             time.sleep(5)
             continue
 
-        recv_bytes += recv_array.nbytes
-        print(f"Rd {i}, received {recv_array.nbytes / 1000000} MB, total {recv_bytes / 1000000} MB")
+        total_bytes += recv_array.nbytes
+        assert recv_data_id == DATA_ID
+        print(
+            f"Rd {i}, received {recv_array.nbytes / 1000000} MB, total {total_bytes / 1000000} MB")
 
-        if recv_bytes >= send_array.nbytes:
+        if total_bytes >= send_array.nbytes:
             break