+++ RELEASE-NOTES.md
@@ -29,7 +29,12 @@ Compared to 0.1.X this version introduces a number of enhancements, although it 
 - Segmenter sync and send stat return tuple (Segmenter::getSyncStats() and Segmenter::getSendStats()) similarly has been replaced with a Segmenter::ReportedStats struct
     - WARNING! This is the point of incompatibility with 0.1.X
 - Segmenter now has a 'warm-up' period of 1 second by default when it sends SYNC packets without allowing data to be sent. This is to help synchronize schedules with the control plane. The length of this period in milliseconds can be set using SegmenterFlags.warmUpMs (also reflected in the INI file format)
+<<<<<<< HEAD
 - Python bindings have been updated to match these changes; python test framework moved to pytest instead of notebooks. Launch Python test suites with `pytest <path> -m unit|b2b`.
+=======
+- Python bindings have been updated to match these changes; python test framework moved to pytest instead of notebooks.
+- Python Reassembler getEvent/recvEvent siganature changes. Now the received data in either numpy or Python bytes is returned by the tuple directly.
+>>>>>>> dev-ejfat6
 
 Note that IP auto-detection features work in simple scenarios where the outgoing interface has at most one address of each type (IPv4 or IPv6) associated with it. The code uses the first address it finds if multiple addresses are found. In the case of e.g. aliased interfaces the results of auto-detection may not be as intended and in this case sender and receiver IP addresses will need to be specified explicitly.